+++ src/arr.js
@@ -107,3 +107,9 @@ export function unique(arr) {
     []
   );
 }
+
+export function concatIfUnique(arr) {
+  return function (smth) {
+    return arr.indexOf(smth) === -1 ? arr.concat(smth) : arr;
+  };
+}