+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - [+] add `indexIn()` and `existsIn()` functions
 - [*] `isEmpty()` function can check object for emptiness
 - [+] add `unique()` function
+- [+] add `concatIfUnique()` function
 
 
 ## 0.1.0 (10.01.2018)